Add line counting, remove inline TODOs
Previous code grew the cell array of strings in the targeted XML file as
they are read in, which is not memory efficient. A subfunction has been
added to count the number of lines in the file in order to accurately
preallocate the cell array in memory. Implementation is currently
Windows specific, but framework is present to handle other OSes.

Removed all inline TODOs in favor of the GitHub issue tracker.

See: #2