Make line counting for unsupported OS less dire
Changes introduced in 4ccb0d1 would cause the entire function to error
out in the event that there is no implemented line counting function to
use for cell array preallocation. This is a bit aggressive and has been
toned down to utilize a growing cell array in memory in the event a line
count cannot be obtained.

See: #2