Fix greedy regex, minor syntax tweaks
Previous regexrep implementation matched removed characters following
any end statement in the XML file. This was fine when the entire XML
file was read into a single character array but is too greedy when each
line of the XML file is read into a cell of a cell array. The regex has
been adjusted to more specifically match the XML footer.

Resolves: #1