Update inline docs, regex fix, syntax tweaks
Update inline documentation to include the addition of user specified
files to process. Minor tweaks to existing comments to align style.

The regex to identify and remove the XML header and footer has been
tweaked slightly to preclude invalid matches in the body of the code
(e.g. if mlapp2classdef() is called in the GUI, mlapp2 is removed). This
issue is also handled by limiting the regex to the first and last lines
of the file, but the regex has also been adjusted in the event more
generic handling is added for the future.

Minor syntax updates to bring additions in line with existing code.