Add classdef parsing for newer MATLAB versions
MATLAB's type function can natively read and display a *.mlapp's class
definition. This is a documented addition to R2016a, but has been tested
successfully in R2014b and newer. This allows us to skip the
intermediate zip file extraction and XML parsing. Though type does not
support an output argument, the output can be piped to a character array
using evalc and exported to an *.m file.

Resolves: #8