Add class property conversion for <R2014b
Starting in R2016a, the documented method to declare class property
types is to specify them after the property declaration with a space in
between (e.g. HeaderLength uint16). This is different from the previous
(undocumented) approach of separating them with the @ symbol (e.g.
HeaderLength@uint16). While the old undocumented approach continues to
function in R2016a, the documented R2016a convention is not backwards
compatible. For m-code loaded in as an array of cells (<R2014b), an
algorithm has been added to locate the property definition blocks and
run a regex to convert all type specifications to the legacy syntax.

See #6, #7