Remove MATLAB version dependency for parsing
Piping the output of type creates a character array of the *.m code,
where the routine to read in an XML file created a cell array of
strings. This introduces compatibility issues and considerations for
some of the regex approaches utilized for function call conversions. To
eliminate the need to consider the differences in input type, the piped
character array from type has been split into a cell array of strings,
mimicking the output of the XML parser.

See #7, #8, f986a72